Updated selectable grid to inherit from HTMLElement
