+++ demo/rhi-ui-selectable-grid.html
@@ -21,10 +21,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
     <meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1, user-scalable=yes">
     <title>Demo :: rhi-ui-selectable-grid :: Rick Hansen Institute</title>
-    <script src="../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
-    <script type="module" src="../node_modules/@polymer/iron-demo-helpers/demo-snippet.js"></script>
-    <script type="module" src="../node_modules/@polymer/marked-element/marked-element.js"></script>
-    <script type="module" src="../packages/selectable-grid/rhi-ui-selectable-grid-demo.js"></script>
+    <script src="../node_modules/markdown-it/dist/markdown-it.min.js"></script>
+    <script type="module" src="../packages/@rhi-ui/selectable-grid/rhi-ui-selectable-grid-demo.js"></script>
     <style>
         body {
             font-family: 'Roboto', 'Noto', sans-serif;
@@ -36,7 +34,7 @@
 </head>
 <body>
     <div class="main-content">
-        <rhi-ui-selectable-grid-demo readmeFile="../packages/selectable-grid/README.md"></rhi-ui-selectable-grid-demo>
+        <rhi-ui-selectable-grid-demo file-uri="../packages/@rhi-ui/selectable-grid/README.md"></rhi-ui-selectable-grid-demo>
     </div>
     <script type="module">
         'use strict';
