+++ packages/@rhi-ui/selectable-grid/package.json
@@ -34,5 +34,5 @@
     "scripts": {
         "prepublishOnly": "tsc -p ../../selectable-grid.tsconfig.json"
     },
-    "version": "1.0.3"
+    "version": "1.0.4"
 }
