+++ tsconfig.json
@@ -22,6 +22,6 @@
     "exclude": [
         "./node_modules",
         "**/*.spec.ts",
-        "./src/selectable-grid/**/*"
+        "./src/@rhi-ui/selectable-grid/**/*"
     ]
 }
